% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_email.R
\name{check_email}
\alias{check_email}
\title{Check for new email on the server}
\usage{
check_email(sid_token, seq)
}
\arguments{
\item{sid_token}{character, session ID token returned from \code{\link{get_email_address}}.}

\item{seq}{integer, the sequence number (id) of the oldest email.}
}
\value{
A list, representing the API response.
}
\description{
Returns a list of the newest messages. The maximum size of the list is 20 items.
}
\details{
The client should not check email too many times as to not overload the server. Do not check if the email expired, the email checking routing should pause if the email expired.
}
\examples{
\dontrun{
library(guerrillamail)
address <- get_email_address()
mail <- check_email(sid_token = address$sid_token, seq = 0)
mail
}
}
\seealso{
\code{\link{https://www.guerrillamail.com/GuerrillaMailAPI.html}}
}
\author{
Philipp Ottolinger
}
