% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_email_address.R
\name{get_email_address}
\alias{get_email_address}
\title{Get an email address from Guerrilla Mail}
\usage{
get_email_address(lang = "en", sid_token = NULL, site = NULL,
  ip = NULL, agent = NULL)
}
\arguments{
\item{lang}{character, a string representing the language code. Currently supported: en, fr, nl, ru, tr, uk, ar, ko, jp, zh, zh-hant, de, es, it, pt.}

\item{sid_token}{character, the session id token used to maintain state.}

\item{site}{character, if you have your own domain and would like to access your custom domains, use this site identifier for your site. Defaults to guerrillamail.com.}

\item{ip}{character, the ip address of the client.}

\item{agent}{character, the user agent of the client.}
}
\value{
A list, representing the API response.
}
\description{
Request a new email address from Guerrilla Mail.
}
\details{
The use of \code{ip} and \code{agent} is recommended when making requests for clients from server side.
}
\examples{
\dontrun{
library(guerrillamail)
address <- get_email_address()
address
}
}
\seealso{
\code{\link{https://www.guerrillamail.com/GuerrillaMailAPI.html}}
}
\author{
Philipp Ottolinger
}
